+++ cli/src/test/kotlin/com/github/zeldigas/text2confl/cli/upload/ContentUploaderTest.kt
@@ -3,6 +3,8 @@ package com.github.zeldigas.text2confl.cli.upload
 import assertk.assertThat
 import assertk.assertions.hasMessage
 import assertk.assertions.isFailure
+import assertk.assertions.isSuccess
+import com.github.zeldigas.confclient.ConfluenceClient
 import com.github.zeldigas.text2confl.convert.Page
 import io.mockk.*
 import io.mockk.impl.annotations.MockK
@@ -15,10 +17,11 @@ import org.junit.jupiter.api.extension.ExtendWith
 
 @ExtendWith(MockKExtension::class)
 internal class ContentUploaderTest(
-    @MockK private val uploadOperations: PageUploadOperations
+    @MockK private val uploadOperations: PageUploadOperations,
+    @MockK private val confluenceClient: ConfluenceClient
 ) {
 
-    private val contentUploader = ContentUploader(uploadOperations)
+    private val contentUploader = ContentUploader(uploadOperations, confluenceClient)
 
     @BeforeEach
     internal fun setUp() {
@@ -71,12 +74,11 @@ internal class ContentUploaderTest(
 
     @Test
     internal fun `Failed upload abort others`() {
-        val fastPage = mockk<Page>() {
-            every { content } returns mockk()
+        val fastPage = aPage("Fast page") {
             every { children } returns emptyList()
         }
-        val failedPage = mockk<Page>()
-        val slowPage = mockk<Page>()
+        val failedPage = aPage("Failed page")
+        val slowPage = aPage("Slow page")
 
         val fastServerPage = ServerPage("fastId", "id", emptyList(), emptyList())
         coEvery { uploadOperations.createOrUpdatePageContent(fastPage, "TEST", "id") } returns fastServerPage
@@ -100,13 +102,62 @@ internal class ContentUploaderTest(
         coVerify(exactly = 0) { uploadOperations.updatePageLabels(slowServerPage, any()) }
     }
 
-    private fun aPage(creationTime: Long, creationTimeRegistry: MutableMap<Page, Long>, block: Page.() -> Unit = {}) : Page {
+    @Test
+    internal fun `Custom parent from attributes is taken into account`() {
+        val defaultPage = aPage("Default page") {
+            every { children } returns emptyList()
+        }
+        val pageWithParentId = aPage("Parent id", mapOf("parentId" to 123, "parent" to "ignored")) {
+            every { children } returns emptyList()
+        }
+        val pageWithParentTitle = aPage("Parent title", mapOf("parent" to "Custom")) {
+            every { children } returns emptyList()
+        }
+
+        val defaultServerPage = ServerPage("defaultId", "id", emptyList(), emptyList())
+        coEvery { uploadOperations.createOrUpdatePageContent(defaultPage, "TEST", "id") } returns defaultServerPage
+
+        val parentIdServerPage = ServerPage("pIdPage", "123", emptyList(), emptyList())
+        coEvery { uploadOperations.createOrUpdatePageContent(pageWithParentId, any(), "123") } returns parentIdServerPage
+
+        val parentTitleServerPage = ServerPage("tIdPage", "345", emptyList(), emptyList())
+        coEvery { uploadOperations.createOrUpdatePageContent(pageWithParentTitle, "TEST", "345") } returns parentTitleServerPage
+
+        coEvery { confluenceClient.getPage("TEST", "Custom") } returns mockk { every { id } returns "345" }
+
+        assertThat {
+            runBlocking { contentUploader.uploadPages(listOf(defaultPage, pageWithParentId, pageWithParentTitle), "TEST", "id") }
+        }.isSuccess()
+
+        coVerify { uploadOperations.createOrUpdatePageContent(defaultPage, "TEST", "id") }
+        coVerify { uploadOperations.createOrUpdatePageContent(pageWithParentId, "TEST", "123") }
+        coVerify { uploadOperations.createOrUpdatePageContent(pageWithParentTitle, "TEST", "345") }
+    }
+
+    private fun aPage(creationTime: Long, creationTimeRegistry: MutableMap<Page, Long>, attributesValues: Map<String, Any?> = emptyMap(), block: Page.() -> Unit = {}) : Page {
         val page = mockk<Page>(block = {
-            every { content } returns mockk()
+            every { content } returns mockk {
+                every { header } returns mockk {
+                    every { attributes } returns attributesValues
+                }
+            }
             every { title } returns "Page created for ${creationTime}ms"
             block()
         })
         creationTimeRegistry[page] = creationTime
+        return page
+    }
+
+    private fun aPage(pageTitle: String, attributesValues: Map<String, Any?> = emptyMap(), block: Page.() -> Unit = {}) : Page {
+        val page = mockk<Page>(block = {
+            every { content } returns mockk {
+                every { header } returns mockk {
+                    every { attributes } returns attributesValues
+                }
+            }
+            every { title } returns pageTitle
+            block()
+        })
         return page
     }
 }