+++ CHANGELOG.md
@@ -7,6 +7,16 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+
+* Ability to override parent via page attributes. This can be handy for use-cases when root pages do not belong to one
+  parent
+
+### Fixed
+
+* Creation of page in confluence cloud - turns out cloud version sets editor property on page creation, so call to
+  create this property fails
+
 ## 0.1.0 - 2022-02-06
 
 Initial release of application.